LayoutTests/crypto/subtle/aes-ctr-import-key-decrypt-expected.txt

CONSOLE MESSAGE: AES-CBC and AES-CTR do not provide authentication by default, and implementing it manually can result in minor, but serious mistakes. We recommended using authenticated encryption like AES-GCM to protect against chosen-ciphertext attacks.
Test decrypting using AES-CTR with an imported 128bit key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bytesToASCIIString(plainText) is expectedPlainText
PASS successfullyParsed is true

TEST COMPLETE
